Worker tests
